photobooth/requirements.txt
fastapi
uvicorn[standard]
websockets
pillow
mediapipe
opencv-python
python-multipart
numpy
